V2.1 - Now usable on Uno and Nano!
V2.1
2015-03-06
- Large scale code clean-up, reduction of variable sizes where they were
unnecessarily large, sharing variables previously unshared between sz
and
rz, and creative use of the send/receive buffer allowed this sketch to
BARELY fit and run with all features enabled on a board with 30K flash
and
2K of RAM.  Uno & Nano users - enjoy.
- Some boards were unstable at baud rates above 9600.  I tracked this
back
to overrunning the SERIAL_TX_BUFFER_SIZE to my surprise.  Added a check
if a flush() is required both in the help and directory listings, as
well
as the sendline() macro.